Moving backend code to different files and adjusting solver function to return string of results.
