+++ train_game_v2/Cargo.toml
@@ -7,3 +7,4 @@ edition = "2021"
 
 [dependencies]
 itertools = "0.10.5"
+serde = { version = "1.0", features = ["derive"] }
